+++ NewNew/SB_DOWN/USER/pid_user.h
@@ -14,10 +14,10 @@ typedef struct
 	int32_t Max_iout;
 }pidTypeDef;
 	
-extern int32_t limit_max(int32_t value,int32_t Max_out);
+extern float limit_max(float value,float Max_out);
 extern void pid_init(pidTypeDef *PID,float p,float i,float d,int32_t Max_out,int32_t Max_iout);
-extern int32_t pid_cal_s(pidTypeDef *PID,float get,float set);
-extern int32_t pid_cal_a(pidTypeDef *PID,float get,float set);
-extern int32_t pid_I_control(pidTypeDef *PID,float get,float set);
+extern float pid_cal_s(pidTypeDef *PID,float get,float set);
+extern float pid_cal_a(pidTypeDef *PID,float get,float set);
+extern float pid_I_control(pidTypeDef *PID,float get,float set);
 
 #endif
